Movi function formatDate para utils
