+++ app/funcionario/horarios/[id]/[asignaturaId]/page.tsx
@@ -11,7 +11,7 @@ import List from '@/components/List/list';
 import Link from 'next/link';
 import { Asignatura } from '@/lib/definitions';
 import { useRouter } from 'next/navigation';
-import { formatDateHour } from '@/utils/utils';
+import { formatDate } from '@/utils/utils';
 
 export default function AsignaturaPage({
   params,
@@ -37,22 +37,6 @@ export default function AsignaturaPage({
     };
     fetch().finally(() => setLoading(false));
   }, [params.asignaturaId]);
-
-  function formatDate(dateString: string): string {
-    const parts = dateString.split('T');
-    if (parts.length !== 2) {
-      throw new Error('Invalid date format. Expected YYYY-MM-DDTHH:mm');
-    }
-
-    const datePart = parts[0];
-    const [year, month, day] = datePart.split('-');
-
-    // Ensure two-digit formatting for month and day
-    const formattedMonth = month.padStart(2, '0');
-    const formattedDay = day.padStart(2, '0');
-
-    return `${formattedDay}/${formattedMonth}/${year}`;
-  }
 
   useEffect(() => {
     const fetch = async () => {
