+++ app/utils/utils.ts
@@ -43,3 +43,20 @@ export const formatDateHour = (dateString: string) => {
   // Return the formatted date string with full year
   return `${formattedMinutes}:${formattedHours} ${formattedDay}/${formattedMonth}/${year}`;
 };
+
+//Para las que vienen formato  yyyy-mm-ddThh:mm con la T en el medio
+export function formatDate(dateString: string): string {
+  const parts = dateString.split('T');
+  if (parts.length !== 2) {
+    throw new Error('Invalid date format. Expected YYYY-MM-DDTHH:mm');
+  }
+
+  const datePart = parts[0];
+  const [year, month, day] = datePart.split('-');
+
+  // Ensure two-digit formatting for month and day
+  const formattedMonth = month.padStart(2, '0');
+  const formattedDay = day.padStart(2, '0');
+
+  return `${formattedDay}/${formattedMonth}/${year}`;
+}
